+++ catalog/app/containers/Bucket/Overview/gql/StatCounts.generated.ts
@@ -23,6 +23,10 @@ export type containers_Bucket_Overview_gql_StatCountsQuery = {
           >
         >
       })
+    | ({ readonly __typename: 'OperationError' } & Pick<
+        Types.OperationError,
+        'name' | 'message' | 'context'
+      >)
   readonly searchObjects:
     | ({ readonly __typename: 'ObjectsSearchResultSet' } & Pick<
         Types.ObjectsSearchResultSet,
@@ -37,6 +41,10 @@ export type containers_Bucket_Overview_gql_StatCountsQuery = {
           >
         >
       })
+    | ({ readonly __typename: 'OperationError' } & Pick<
+        Types.OperationError,
+        'name' | 'message' | 'context'
+      >)
 }
 
 export const containers_Bucket_Overview_gql_StatCountsDocument = {
@@ -114,6 +122,21 @@ export const containers_Bucket_Overview_gql_StatCountsDocument = {
                     ],
                   },
                 },
+                {
+                  kind: 'InlineFragment',
+                  typeCondition: {
+                    kind: 'NamedType',
+                    name: { kind: 'Name', value: 'OperationError' },
+                  },
+                  selectionSet: {
+                    kind: 'SelectionSet',
+                    selections: [
+                      { kind: 'Field', name: { kind: 'Name', value: 'name' } },
+                      { kind: 'Field', name: { kind: 'Name', value: 'message' } },
+                      { kind: 'Field', name: { kind: 'Name', value: 'context' } },
+                    ],
+                  },
+                },
               ],
             },
           },
@@ -166,6 +189,21 @@ export const containers_Bucket_Overview_gql_StatCountsDocument = {
                           ],
                         },
                       },
+                    ],
+                  },
+                },
+                {
+                  kind: 'InlineFragment',
+                  typeCondition: {
+                    kind: 'NamedType',
+                    name: { kind: 'Name', value: 'OperationError' },
+                  },
+                  selectionSet: {
+                    kind: 'SelectionSet',
+                    selections: [
+                      { kind: 'Field', name: { kind: 'Name', value: 'name' } },
+                      { kind: 'Field', name: { kind: 'Name', value: 'message' } },
+                      { kind: 'Field', name: { kind: 'Name', value: 'context' } },
                     ],
                   },
                 },