+++ catalog/app/containers/Search/Sort.tsx
@@ -78,6 +78,7 @@ export default function Sort({ className }: SortProps) {
       switch (r.__typename) {
         case 'EmptySearchResultSet':
         case 'InvalidInput':
+        case 'OperationError':
           return false
         case 'ObjectsSearchResultSet':
         case 'PackagesSearchResultSet':
@@ -100,6 +101,7 @@ export default function Sort({ className }: SortProps) {
       value={value}
       onChange={handleChange}
       ButtonProps={{ classes: buttonClasses, size: 'medium' }}
+      shrink={sm}
     >
       {sm ? <M.Icon>sort</M.Icon> : 'Sort by:'}
     </SelectDropdown>